feedback from designer discussion
