+++ src/WebUI/Types.d.ts
@@ -90,26 +90,3 @@ export interface IVssComponentProperties extends IBaseProps {
     [property: string]: IObservable<any> | any;
 
 }
-
-/**
- * https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-phase
- **/
-export enum PodPhase {
-    Pending = "Pending",
-    Running = "Running",
-    Succeeded = "Succeeded",
-    Failed = "Failed",
-    Unknown = "Unknown",
-    Completed = "Completed",
-    CrashLoopBackOff = "CrashLoopBackOff",
-}
-
-export const PodPhaseToStatus: { [index: string]: IStatusProps } = {
-    "Running": Statuses.Success,
-    "Succeeded": Statuses.Success,
-    "Completed": Statuses.Success,
-    "Pending": Statuses.Failed,
-    "Failed": Statuses.Failed,
-    "Unknown": Statuses.Queued,
-    "CrashLoopBackOff": Statuses.Failed,
-}
