+++ src/WebUI/WorkLoads/WorkloadsFilterBar.tsx
@@ -20,7 +20,7 @@ export interface IWorkloadsFilterBarProps extends IVssComponentProperties {
 export class WorkloadsFilterBar extends BaseComponent<IWorkloadsFilterBarProps> {
     public render(): React.ReactNode {
         return (<KubeFilterBar filter={this.props.filter}
-            keywordPlaceHolder={Resources.WorkloadsFilterText}
+            keywordPlaceHolder={Resources.WorkloadText.toLocaleLowerCase()} // lowercase to show in filter
             pickListPlaceHolder={Resources.KindText}
             pickListItemsFn={this._pickListItems}
             listItemsFn={this._listItems}
